+++ server/routes/index.js
@@ -1,9 +1,11 @@
 const { Router } = require('express');
 const router = Router();
-const rideController = require('../controllers/Controller1.js')
+const rideController = require('../controllers/Ride.js')
+const reviewController = require('../controllers/Review')
 
 
-// router.get('/', (req, res) => res.send('This is root!'))
 router.get('/', rideController.getAllRides)
+router.get('/reviews/:id', reviewController.getAllReviews)
+router.post('/reviews/new', reviewController.createReview)
 
 module.exports = router;
